50 test fonctionnel + Uniformité du code + Correction de plein de fon…

+++ CommuMoney/CommuMoney.METIER/Metier/Projet_METIER.cs
@@ -13,12 +13,12 @@ namespace CommuMoney.METIER.Metier
         public int ID { get; set; }
         public string Nom { get; set; }
         public int ID_Personne { get; set; }
-        public float Total_Montant { get; set; }
-        public float Moyenne { get; set; }
+        public double Total_Montant { get; set; }
+        public double Moyenne { get; set; }
         public DateTime? Date_Soiree { get; set; }
 
-        public Projet_METIER(string nom, int id_personne, float total_montant, float moyenne, DateTime? date_soiree) => (Nom, ID_Personne, Total_Montant, Moyenne, Date_Soiree) = (nom, id_personne, total_montant, moyenne, date_soiree);
-        public Projet_METIER(int id, string nom, int id_personne, float total_montant, float moyenne, DateTime? date_soiree) => (ID, Nom, ID_Personne, Total_Montant, Moyenne, Date_Soiree) = (id, nom, id_personne, total_montant, moyenne, date_soiree);
+        public Projet_METIER(string nom, int id_personne, double total_montant, double moyenne, DateTime? date_soiree) => (Nom, ID_Personne, Total_Montant, Moyenne, Date_Soiree) = (nom, id_personne, total_montant, moyenne, date_soiree);
+        public Projet_METIER(int id, string nom, int id_personne, double total_montant, double moyenne, DateTime? date_soiree) => (ID, Nom, ID_Personne, Total_Montant, Moyenne, Date_Soiree) = (id, nom, id_personne, total_montant, moyenne, date_soiree);
 
         #region Insert
         public void Insert()
